feat: organize enums and components base name
