+++ packages/asset-packs/package.json
@@ -17,7 +17,7 @@
     "@aws-sdk/client-s3": "^3.405.0",
     "@aws-sdk/lib-storage": "^3.405.0",
     "@dcl/hashing": "^3.0.4",
-    "@dcl/js-runtime": "7.18.1",
+    "@dcl/js-runtime": "7.14.0",
     "@dcl/react-ecs": "7.15.2",
     "@dcl/sdk-commands": "7.18.1",
     "@types/mime-types": "^2.1.1",
