+++ packages/asset-packs/src/enums.ts
@@ -1,6 +1,21 @@
 import type { YGAlign, YGJustify } from '@dcl/ecs';
 import { getLatestVersionName } from './versioning';
 
+export const BaseComponentNames = {
+  ACTION_TYPES: 'asset-packs::ActionTypes',
+  ACTIONS: 'asset-packs::Actions',
+  COUNTER: 'asset-packs::Counter',
+  TRIGGERS: 'asset-packs::Triggers',
+  STATES: 'asset-packs::States',
+  COUNTER_BAR: 'asset-packs::CounterBar',
+  ADMIN_TOOLS: 'asset-packs::AdminTools',
+  VIDEO_SCREEN: 'asset-packs::VideoScreen',
+  REWARDS: 'asset-packs::Rewards',
+  TEXT_ANNOUNCEMENTS: 'asset-packs::TextAnnouncements',
+  VIDEO_CONTROL_STATE: 'asset-packs::VideoControlState',
+  SCRIPT: 'asset-packs::Script',
+} as const;
+
 export const ComponentName = {
   ACTION_TYPES: 'asset-packs::ActionTypes',
   ACTIONS: 'asset-packs::Actions',
@@ -112,7 +127,7 @@ export enum ActionType {
 }
 
 // Re-export trigger enums from versioning (source of truth)
-export { TriggerType, TriggerConditionType, TriggerConditionOperation } from './versioning';
+export { TriggerType, TriggerConditionType, TriggerConditionOperation } from './trigger-enums';
 
 // Defined values instead of using from @dcl/ecs because Schemas doesn't support const enums
 export enum AlignMode {
