+++ packages/asset-packs/src/versioning/definitions/index.ts
@@ -1,3 +1,2 @@
 export * from './counter';
 export * from './triggers';
-export * from './trigger-enums';
